fix types in messageport
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/branches/gnustep_testplant_branch@40159 72102866-910b-0410-8b05-ffd578937521

+++ Source/win32/NSMessagePort.m
@@ -966,10 +966,10 @@ again:
 }
 
 - (BOOL) sendBeforeDate: (NSDate*)when
-		  msgid: (int)msgId
+		  msgid: (NSInteger)msgId
              components: (NSMutableArray*)components
                    from: (NSPort*)receivingPort
-               reserved: (unsigned)length
+               reserved: (NSUInteger)length
 {
   NSMutableData		*h = nil;
   NSMutableData		*first;
